+++ examples/pub_sub.js
@@ -4,7 +4,7 @@ var redis = require('redis'),
     client1 = redis.createClient(), msg_count = 0,
     client2 = redis.createClient();
 
-// Most clients probably don't do much on 'subscribe'.  This example uses it to coordinate things within one program.
+// Most clients probably don't do much on 'subscribe'. This example uses it to coordinate things within one program.
 client1.on('subscribe', function (channel, count) {
     console.log('client1 subscribed to ' + channel + ', ' + count + ' total subscriptions');
     if (count === 2) {