+++ test/commands/keys.spec.js
@@ -15,8 +15,6 @@ describe("The 'keys' method", function () {
 
             beforeEach(function (done) {
                 args = args || {};
-                // This is going to test if the high water is also respected
-                args.command_queue_high_water = 100;
                 client = redis.createClient.apply(redis.createClient, args);
                 client.once("ready", function () {
                     client.flushdb(done);